+++ NEWS.md
@@ -1,7 +1,7 @@
 # hubValidations 0.0.0.9007
 
 * `validate_pr()` now has arguments for controlling modification/deletions check are performed on model output and model metadata files (#65).
-  - `file_modify_check`, which controls whether modification/deletion checks are performed and what is retuned if modifications/deletions detected.
+  - `file_modification_check`, which controls whether modification/deletion checks are performed and what is returned if modifications/deletions detected.
   - `allow_submit_window_mods`, which controls whether modifications/deletions of model output files are allowed within their submission windows.